flattened cli
